requirements.txt
gevent==1.3.4
greenlet==0.4.13
locustio==0.9.0
requests==2.18.4
protobuf==3.5.2
